pyproject.toml
[tool.poetry]
name = "fastapi_benchmark"
version = "0.1.0"
description = ""
authors = ["andrpocc <andrey.pochatkov@mail.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
tortoise-orm = "^0.19.2"
aerich = "^0.6.3"
asyncodbc = "^0.1.1"
asyncpg = "^0.26.0"
python-dotenv = "^0.20.0"
uvicorn = "^0.18.2"
fastapi-utils = "^0.2.1"
requests = "^2.28.1"
orjson = "^3.7.12"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
pylint = "^2.14.5"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
